requirements.txt
ultralytics
pillow
tk
pandas
openpyxl
opencv-python-headless